fix build dependencies in Dockerfile.raspberry
The base image for the Raspberry don't have zlib1g-dev or the
Python wheel (python-wheel and python3-wheel) installed.

The first results in a wget-lua that doesn't have zlib support, which
cause an exception upon upload, with an invalid data package uploaded
to the server.

The second one doesn't appear to cause any problems, but it doesn't
look pretty when building the image.