Scope watchtower to just this warrior
The example docker-compose includes a watchtower service. As configured this service would update any containers running on this docker host with the right label. This might be surprising to some people who don't already use watchtower (but may have copy/pasted the watchtower label at some point when reading tutorials/examples).

The fix is to scope the watchtower service in the example docker-compose.yml to just this stack.